Fixed integer overflow bug in pixel width/height

+++ src/constants.rs
@@ -6,8 +6,8 @@ pub const WINDOW_HEIGHT: u32 = 448;
 pub const DISPLAY_WIDTH: u8 = 64;
 pub const DISPLAY_HEIGHT: u8 = 32;
 
-pub const PIXEL_WIDTH: u8 = WINDOW_WIDTH as u8 / DISPLAY_WIDTH;
-pub const PIXEL_HEIGHT: u8 = WINDOW_HEIGHT as u8 / DISPLAY_HEIGHT;
+pub const PIXEL_WIDTH: u8 = (WINDOW_WIDTH / DISPLAY_WIDTH as u32) as u8;
+pub const PIXEL_HEIGHT: u8 = (WINDOW_HEIGHT / DISPLAY_HEIGHT as u32) as u8;
 
 pub const PIXEL_COLOR: &Color = &Color::RGB(255, 255, 255);
 pub const BACKGROUND_COLOR: &Color = &Color::RGB(0, 0, 0);
